+++ src/api/index.js
@@ -109,14 +109,7 @@ export const reorderTracksInPlaylist = (playlistId, start, end) => {
 
     return fetch(URL, {
         ...corsParams
-    }).then(
-        response => response.json(),
-        error =>
-            console.error(
-                'Error fetching playlist reorder tracks operation',
-                error
-            )
-    );
+    }).then(response => response.json(), error => error);
 };
 
 export const getLogOutUser = () => {