+++ src/components/MyPlaylists/index.js
@@ -53,8 +53,11 @@ let scroll = Scroll.animateScroll;
 
 class MyPlaylists extends PureComponent {
     static propTypes = {
+        startPlaylistTracksReorderProcess: PropTypes.func.isRequired,
         myPlaylistsHasOpenPlaylist: PropTypes.bool.isRequired,
         removePlaylistFromFinal: PropTypes.func.isRequired,
+        reorderPlaylistTracks: PropTypes.func.isRequired,
+        setPlaylistDragStatus: PropTypes.func.isRequired,
         addPlaylistToFinal: PropTypes.func.isRequired,
         fetchMyPlaylists: PropTypes.func.isRequired,
         myPlaylists: PLAYLISTS_PROPTYPE.isRequired,
@@ -177,6 +180,16 @@ class MyPlaylists extends PureComponent {
         this.searchInputRef.focus();
     };
 
+    _handlePlaylistTracksReorder = (playlistId, trackId, startPos, endPos) => {
+        const { startPlaylistTracksReorderProcess } = this.props;
+        startPlaylistTracksReorderProcess(
+            playlistId,
+            trackId,
+            startPos,
+            endPos
+        );
+    };
+
     componentDidMount() {
         toTop();
 
@@ -253,14 +266,6 @@ class MyPlaylists extends PureComponent {
             </div>
         );
 
-        const ListOfMyPlaylists = (
-            <List
-                onClickMain={this._handleAddPlaylist}
-                onClickItem={this._handleClickPlaylist}
-                items={playlists}
-                isPlaylist={true}
-            />
-        );
         const playlistCounter = (
             <div className={classes.playlistRemaining}>
                 {playlistsRemaining}
@@ -300,7 +305,13 @@ class MyPlaylists extends PureComponent {
                             this._handleCanScrollUp(false);
                         }}
                     />
-                    {ListOfMyPlaylists}
+                    <List
+                        onClickMain={this._handleAddPlaylist}
+                        onClickItem={this._handleClickPlaylist}
+                        items={playlists}
+                        isPlaylist={true}
+                        onDragAndDrop={this._handlePlaylistTracksReorder}
+                    />
                     <Waypoint
                         onEnter={() => {
                             this._handleCanScrollUp(true);