+++ src/connectPage.js
@@ -2,6 +2,7 @@ import { connect } from 'react-redux';
 import { push } from 'react-router-redux';
 import * as R from 'ramda';
 import {
+    startPlaylistTracksReorderProcess,
     setComponoformAddExistingStatus,
     removePlaylistTrackFromFinal,
     setOpenStatusPublicPlaylists,
@@ -19,11 +20,13 @@ import {
     setFinalPlaylistPublic,
     searchPublicPlaylists,
     setPublicPlaylistOpen,
+    reorderPlaylistTracks,
     checkIfAuthenticated,
     setMyPlaylistVisited,
     clearComponoformData,
     setFinalPlaylistOpen,
     launchPlaylistMerger,
+    setPlaylistDragStatus,
     setFinalPlaylistUrl,
     setSelectedPlaylist,
     fetchPlaylistTracks,
@@ -33,6 +36,7 @@ import {
     setFinalSearchTerm,
     addPlaylistToFinal,
     setComponofyMode,
+    startDragAndDrop,
     fetchMyPlaylists,
     setPlaylistOpen,
     setMySearchTerm,
@@ -309,6 +313,25 @@ export const mapDispatchToProps = dispatch => ({
 
     setCurrentSelectionOffset(offset) {
         dispatch(setCurrentSelectionOffset(offset));
+    },
+
+    reorderPlaylistTracks(playlistId, trackId, startPos, endPos) {
+        dispatch(reorderPlaylistTracks(playlistId, trackId, startPos, endPos));
+    },
+
+    setPlaylistDragStatus(playlistId, hasReorderRequest) {
+        dispatch(setPlaylistDragStatus(playlistId, hasReorderRequest));
+    },
+
+    startPlaylistTracksReorderProcess(playlistId, trackId, startPos, endPos) {
+        dispatch(
+            startPlaylistTracksReorderProcess(
+                playlistId,
+                trackId,
+                startPos,
+                endPos
+            )
+        );
     }
 });