+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANEGLOG
 
+### v1.1.0
+
+* Major feature: ability to reorder tracks in a playlist owned by the user
+* The rest of thea features can be found in [v1.1.0 Feature Board](https://github.com/DalerAsrorov/componofy/projects/4).
+
 ### v1.0.1
 
 * Major feature: ability to merge public/private/personal playlists into an existing user's playlist.